requirements.txt
alembic==1.15.2
async-timeout==5.0.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==42.0.5
Deprecated==1.2.18
dnspython==2.7.0
dominate==2.9.1
email_validator==2.1.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.1
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SeaSurf==2.0.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
Mako==1.3.9
Flask-Mail==0.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
python-dotenv==1.0.1
redis==5.2.1
rich==13.9.4
sentry-sdk==2.25.1
SQLAlchemy==2.0.36
typing_extensions==4.13.1
urllib3==2.3.0
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zipp==3.21.0
